requirements.txt
flask
paho-mqtt
pandas
python-dotenv
gunicorn
sqlalchemy
pymysql
